:art: Abstract token validation from confirmEmail
