+++ resources/auth/middlewares.js
@@ -1,3 +1,5 @@
+const validateToken = require('../../utils/validateToken');
+const { EMAIL_SECRET } = require('../../config');
 const model = require('./model');
 
 exports.checkUserExists = async (req, res, next) => {
@@ -17,5 +19,25 @@ exports.checkEmailExists = async (req, res, next) => {
     res.status(404).json({ message: 'User with this email does not exists' });
   } else {
     next();
+  }
+};
+
+exports.validateToken = async (req, res, next) => {
+  try {
+    const { token } = req.body;
+    const decodedToken = validateToken(token, EMAIL_SECRET);
+
+    const userId = decodedToken.subject;
+    const response = await model.confirmEmail(userId);
+
+    if (response) {
+      req.userId = decodedToken.subject;
+      req.userEmail = response.email;
+      next();
+    } else {
+      res.status(400).json({ message: `Email confirmation failed!` });
+    }
+  } catch (error) {
+    res.status(400).json({ message: `Confirmation failed: ${error.message}!` });
   }
 };
